extracted math into LaneSwitchUtils
